requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.1.1
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
setuptools==75.1.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
qrcode~=8.0
Faker~=30.1.0